Create automation 0.6.0.py
Sure, here is a possible implementation of the add_cell_automation method that creates cells in the location using the arrow keys and default name and description:

Note that this implementation assumes that the opposite_direction method from the Cell class is already implemented. Also, this implementation does not include any error handling for unexpected keyboard input, so you may want to add that as well.