Create map gen 0.5.py
Here, LocationMap takes in a dictionary of Location objects, and creates a map as a string using the conventions you described. The map is created by iterating over each location in the dictionary and adding its cells to the map. The get_connected_locations() method of Location is used to determine if a cell is connected to another location, and if so, an "x" is added to the corresponding side of the cell in the map.