Create map gen 1.0.py
Yes, it is possible to create a map of all locations connected to each other. One approach is to use graph theory to represent each location as a node in a graph and the connections between locations as edges. Then, you can use a graph traversal algorithm, such as depth-first search or breadth-first search, to visit each node in the graph and create a map of the entire network of connected locations.

Here's an example implementation:


This LocationMap class takes in a dictionary of all the locations and their connections, and builds a graph representation of the locations using a defaultdict of lists. It then uses a depth-first search algorithm to visit each node in the graph and create a map of the entire network of connected locations.

The traverse_map method calls the traverse method for each unvisited node in the graph and returns a list of all the maps of connected locations.

The traverse method takes in a location ID and recursively visits all its connections to build a map of the connected locations. It uses the walls attribute of each cell in a location to determine whether to add a horizontal or vertical line to the map.

Note that the map generated by this class is not interactive, but rather a static representation of the network of connected locations.